refactor(LobbyScreen):exported PlayerList to external module
